refactor things in server and environment
