+++ spec/dummy/config/webpack/test.js
@@ -4,6 +4,29 @@ const environment = require('./environment')
 const serverConfig = require('./server')
 const merge = require("webpack-merge")
 
+
+//adding exposeLoader
+const exposeLoader = {
+    test: require.resolve('jquery'),
+    use: [ { loader: 'expose-loader', options: 'jQuery' } ]
+}
+environment.loaders.insert('expose', exposeLoader, { after: 'file'} )
+
+//adding es5Loader
+const es5Loader = {
+    test: require.resolve('react'),
+    use: [ { loader: 'imports-loader', options: { shim: 'es5-shim/es5-shim', sham: 'es5-shim/es5-sham' } } ]
+}
+environment.loaders.insert('react', es5Loader, { after: 'sass'} )
+
+//adding jqueryUjsLoader
+const jqueryUjsLoader = {
+    test: require.resolve('jquery-ujs'),
+    use: [ { loader: 'imports-loader', options: { jQuery: 'jquery' } } ]
+}
+environment.loaders.insert('jquery-ujs', jqueryUjsLoader, { after: 'react'} )
+
+
 const clientConfig = merge(environment.toWebpackConfig(), {
     mode: 'development',
     entry: {
